requirements.txt
wheel==0.36.2
Werkzeug==1.0.1
SQLAlchemy==1.4.27
six==1.16.0
setuptools==57.0.0
python-dotenv==0.15.0
python-dateutil==2.8.2
pip==21.1.2
MarkupSafe=2.0.1
Jinja2==3.0.3
itsdangerous==2.0.1
greenlet==1.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-RBAC==0.5.0
Flask-HTTPAuth==4.2.0
Flask-Cors==3.0.10
eventlet==0.30.0
dnspython==1.16.0
click==7.1.2


